--- get_profile/templates/index.html
@@ -1,3 +1,0 @@
-<h1> {{ all_data }} </h1>
-<h1> {{ username }} </h1>
-<h1> {{ type }} </h1>
